Implement condensed channel lineup API (#59)
Now we return a lineup list with a similar schema to what is saved in
the DB, along with a materialized program lookup map. Materialized
program details are much larger in size than lineup lists, so this
should make things a bit snapper and help support really 'large' lineups
(i.e. Channels with slot schedules that are precalculated for a year+)

This also contains some ideas of how to streamline tv guide generation
and potentially be able to do it on the fly.

Lastly, there is a lot of code cleanup. I verified that both the web and
server packages build successfully.