Standardize on plex key that excludes '/library/metadata' prefix. Fixes #268
Initially we were trying to mimic the dizqueTV data model. Programs in
DTV saved both Plex's 'ratingKey' and 'key' values. These are
effectively the same thing, from what I can tell, with the 'key' just
being a 'ratingKey' prefixed by '/library/metadata/'.

The issue in this bug was that on the frontend, we were creating
'uniqueId' using the 'ratingKey' and on the backend, we were creating it
using the 'key'. We would end up with LineupPrograms without a proper DB
Program reference, which would break various things on the save path.

This change does a few things:
1. Standardizes on Plex ratingKey as the 'externalKey' on a Program
2. Deprecates the plexRatingKey field on the Program, since eventually
   these will not be Plex-specific anyway and it just causes more
   confusion
3. Updates references to the aforementioned fields

Next steps:
1. Use branded types for additional string type safety - we have a lot
   of srtings flying around!