Fixes nested InlineModals which were passing the wrong childItemGuid
Kudos for the key insight from @markdavella here -- sibline
InlineModal/PlexGridItems do **not** have any relation to one another,
in terms of which item they are 'pointing' to.

This meant that we were binding the wrong child item GUID to each
InlineModal. In fact, we have to render InlineModals at each index to
allow it to be injected at any point within the grid. HOWEVER, they
should always point to the _open_ child item GUID, which is controlled
by the parent element.

Testing:

Opened TV Collection > Show > Season > Episodes -- a complete hierarchy.