Automatically fix duration mismatches between DB and lineup
A consequence of decoupling lineups from the DB items is that sometimes
things can get out of sync. In this case, I experienced the following
scenario:

1. Program 1 added to the DB with duration D
2. Program 1 was added to Channel A and Custom Show X
3. Outside of Tunarr, a new version of this Program was imported to my
   media server, which had duration D-1
4. Channel A was updated, and with it, associated program metadata was
   updated from the media server source. In the DB, Program 1 now had
duration D-1. Channel A's lineup was also correctly updated
5. Channel B, which uses Custom Show X in its lineup (containing Program
   1) did _not_ have its lineup updated, because it was not aware that
      Program 1 had changed in the DB.

The solution here is twofold:

1. In guide building, account for scenarios where the item's duration in
   the DB is _less than_ the reported duration in the lineup file by
adding flex. We've also added logging to capture the opposite scenario
(DB time is longer than lineup time)
2. Add a new task that reconciles durations. This fires on a schedule
   (for now) and also after the main path where programs are
inserted/updated in the DB.