Implement program_external_id table
Motiviation:
Adding this table is a bit of a future-proofing mechanism. We want to
decouple the notion of a 'program' from its external references.
Currently, the only external reference Tunarr knows of is 'plex'.
Eventually, we will add different sources for programming. In theory, a
program could exist in multiple sources. And not only in multiple media
sources, but on other sites that contain more metadata, such as IMDB and
TMDB. The new DB table generalizes Program to have a 1:M relationship
with external references. Additionally, this allows us to maintain
stable and _unstable_ IDs for various sources (ex. plex rating key)
without losing metadata sourcing for a Program.

Changes:
* Migration to add `program_external_id` table
* Backfill fixer task to fill in details from existing programs in the
  DB
* Changes to program DB save path to save to new table and to start
  saving Plex GUID
* Self-heal when Plex rating key changes. This code is triggered when
  starting a stream for an item. If we have a Plex GUID available, we
will attempt to heal the item in the DB and start the stream
* Use the new table in the stream item calculator

TODO:
* Start saving Plex GUID for program_groupings, which have a separate
  external key table
* Start saving other external IDs for programs, such as IMDB ID and TMDB
  ID