Implement 'select all' functionality in program selector (#482)
Closes #481

This adds new functionality to 'select all' of the visible items in the
program selector. The use-case here is to execute a Plex search/filter
and then add all of the results at once

Also includes the following changes:
* Refactor giant plex hooks file into multiple smaller files
* Breakout some common plex search queries into new hooks
* Use FixedList in the selected programming list drawer - I noticed a
  slowdown after implementing select all; this helps.
* Refactor the way the SelectedProgrammingList component works a bit.
  We're no longer mounting this in App with hard-coded paths -- there's
more fixes here, though. Now we have a temporary drawer with various
toggle states, etc etc. This should free up the viewport to view content
* Fixes a bug in Plex search/filtering for TV shows
* Enables the use of path aliases in imports using "@/", since we're
  starting to deal with a lot of files in many nested directories. This
cleans things up a bit