Removes swagger-ui Fastify plugin, hopefully temporarily
We've seen reports like
https://github.com/chrisbenincasa/tunarr/issues/550 where swagger-ui
prevents the server from starting, permanently. I haven't been able to
reproduce this yet. We've opened an issue on the plugin repo to see if
they have ideas:
https://github.com/fastify/fastify-swagger-ui/issues/156

For now, we're not really in need of this anyway, so just remove it.

I've also updated a bunch of deps and shimmed in the "DOM" types in our
tsconfig since there are lots of errors when compiling during the lib
check phase for things like esbuild, vitest.