fix: workaround for macOS binary release
Fixes #599

Due to multiple issues surrounding single executable generation for
macOS, including issues with code signing due to executable mangling
(necessary when using Nexe with binary resources, i.e. the Tunarr
webapp), it seems the simplest solution is to simply bundle the relevant
version of nodejs itself along with a simple script to run Tunarr.

This essentially mimics the inside of our Docker containers and, in a
way, how the bundled executable works too... it's a little gnarly, but
it seems to work.