fix: resolve Jellyfin items when underlying file changes (#933)
Jellyfin item IDs are not stable. When the underlying file changes in
Jellyfin, apparently even if this file has the same path, Jellyfin will
regenerate the ID. An example of this is when an upgrade occurs to a
file from an *arr suite program. We can use existing IMDB/TMDB/TVDB IDs
we found when initially saving the item to search Jellyfin and find its
replacement. This is done on-the-fly when attempting to stream an item
that is subsequently not found.