feat: add "transcode configurations"
This adds the notion of creating and maintaining transcode
configurations. This is a generalization of the current ffmpeg settings
flow. Users can edit various settings relating to transcoding in each
config and then assign a configuration to a channel. This allows for
more fine-grained customization of transcoding per-channel and also
allows for users to test new configurations without affecting the global
state of Tunarr.