fix: properly handle redirects in scheduler
This fixes a bug where the scheduler thought it could progress to the
next program in the schedule, since it assumed the the previously
scheduled item matched the full duration of the lineup slot. This isn't
necessarily true in situations like redirects, which defer their program
choices to the target channel and may need several loops to fill the
whole source channel lineup's "slot".