feat: implement existing program tools using random slots
This change begins the concept of creating schedules based on existing
programming tools using the slot schedulers. The first of these is
recreating "cyclic shuffle" + "tweak weights" functionality directly in
the random slot scheduler. To do this, the following functionality was
added to the random slot scheduler:

1. Ability to create slots based on absolute # of programs per slot,
   rather than a fixed duration
2. A new program orderer called "ordered shuffle". This is effectively
   the "cyclic shuffle' functionality. Namely, chronological (or index
   based for custom shows) ordering with a random start point

The concept is called "presets" - this allows users to generate a
schedule that would normally require tinkering with program tools. The
idea is to bake in various "presets" based on these existing tools, plus
other convenience functions, to ease channel creation.