feat: improvements to random slot weighting
Includes the ability to keep weights locked to one another - or by
default, keep them "unlocked" (the classic DTV experience). This is
mainly a UX feature, but serves as the basis for implementing certain
programming tools as random slots (coming in a follow-up). Additionally,
we show actual weights of generated schedules in a tooltip.