fix: allow scheduling custom shows that contain the same programs
This was an oversight in how we de-duplicate programs before save. We
need to de-dupe based on the (custom show, program id) pair for custom
programs. Additionally, this change adds the ability to send "persisted"
programs as part of the lineup update request, removing the need to send
full program details for every program in the lineup.