feat: show version mismatch warning on frontend
If the backend updates without the user reloading their Tunarr tab, the
loaded code in the browser could skew from the backend. This could lead
to errors and other issues due to not promise of cross-version of
compatibility between Tunarr backend and frontend. This change shows a
prompt for the user to reload their browser if the version baked into
the frontend assets doesn't match the version received from the server.