feat: support in-order slot scheduling
This feature changes "Random Slots" to simply "Slot Editor"; the slot
editor now allows slots to be scheduled in the order they are defined,
rather than chosen at random at scheduling time. This allows for more
fine-tuned scheduling reminiscent of "block shuffle", but more powerful.

Additionally, this adds unused components to display schedules in a
monthly and weekly calendar view. These are unfinished and not exposed
anywhere.

Lastly, there are a few bug fixes to the slot editor scheduler and UI
included.