feat!: expose advanced transcode parameters
This feature is mainly to provide a migration path from legacy transcode
pipeline users to the new pipeline in preparation of the former's
deprecation and removal. In general, users should not need to use these
first set of parameters (disabling various hardware accelerated
features) as doing so implies a bug that should be fixed. However, this
should provide a more smooth path for legacy holdouts to get onto the
new pipeline and allow them to experiment with the full, new pipeline,
without potentially wrecking their existing setup (via multiple
transcode configs).

Closes #1090