fix: allow selecting parent Jellyfin items in list view
This began after removing all userIds from Jellyfin API queries and
using only a token. Jellyfin doesn't seem to return ChildCount when
requested in this way, because it is a user-scoped field (based on user
permissions). We have to make a judgement based on the item type that
there are children under these parents to allow for selection