feat: allow scheduling shows that are not part of channel lineup
Previously slot schedulers held the restriction that shows must first be
added to the channel's lineup (and saved) before they could be used in
slots. This also had the detriment that programs would be _removed_ if
they were not included in the generated schedule. Or, a user could make
the mistake of not adding all of the show they wanted to the channel and
then become confused when the generated schedule didn't include certain
content.

This change fixes all of these issues. It fully decouples picking shows
for slots from the state of a channel's lineup, now that we have the
capability to do full "search". Now, creating slot schedules is nearly
fully decoupled from channel lineup content (save for movies) and users
are not required to visit the "add programming" flow at all and can
instead just start building their schedules.