fix: remove Plex auto-channel updater
This "feature" has been the source of confusion and is under-developed,
potentially causing damage to users' Plex Live TV setup by overwriting
channel mappings. It is unclear what you are getting into when enabling
this feature. It is safer to remove this then deal with any fallout
relating to its "under developed" nature and Plex Live TV's
shortcomings.