fix: more balanced cyclic shuffle implementation
Creates a more balanced implementation of cyclic shuffle. The
previous implementation was overly "normal" and would create
a long tail of shows which had more episodes than other shows
in the lineup. The new implementation shuffles the constituent
episodes first for a more "random" feel but then restores the
episode ordering to remain true to cyclic shuffle.